requirements.txt
streamlit
google-generativeai
python-dotenv
ipython
langchain
PyPDF2
chromadb